requirements.txt
albumentations==2.0.8
faiss_gpu==1.7.2
imgaug==0.4.0
matplotlib==3.10.7
numpy==2.3.5
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==12.0.0
PyYAML==6.0.2
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
timm==1.0.22
torch==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
transformers==4.52.3
